requirements.txt
aiohttp
asyncio
motor
aiodns
python-dotenv